+++ packages/core-ui/client/helpers.js
@@ -35,6 +35,12 @@ Template.registerHelper('routeParamIs', function(name, val) {
 	var cur = Router.current();
 	return cur.getParams()[name] === val;
 });
+Template.registerHelper('routeBack', function() {
+	var cur = Router.current(),
+		u = (cur && cur.url)
+		p = u.split('/').slice(0,-1).join('/');
+	return p || 'javascript:history.back()';
+});
 
 
 Template.registerHelper('connectionStatus', function() {