+++ packages/core-ui/client/views/formLogin.js
@@ -18,4 +18,4 @@ Template.formLogin.helpers({
 		}
 		return c;
 	}
-})
+});